feat: add activity highlight toggle and fix usage tracking accuracy
- Add ToggleSwitch to enable/disable activity highlighting on devices page
- Fix ActiveSeconds calculation: use timestamps instead of counting input events
- Track last active time per device for accurate time measurement
- Only count time intervals < 1 second to prevent counting gaps when user is away
- Change ActiveSeconds from long to double for millisecond precision

+++ Models/DeviceUsageRecord.cs
@@ -9,7 +9,7 @@ public class DeviceUsageRecord
 {
     public string DeviceId { get; set; } = "";
     public DateTime Date { get; set; }
-    public long ActiveSeconds { get; set; }
+    public double ActiveSeconds { get; set; }
     public long EnabledSeconds { get; set; }
 
     /// <summary>
